Add support to allow binding a function to a "replace key"
add support to allow binding a lisp function to a "replace key" in
`speck-replace-keys' and/or `speck-replace-map', so that one can define a
function to do more hacking. for instance, one can define a function, and bind
it to the esc key in `speck-replace-keys' and/or `speck-replace-map', to decide
whether to accept the speck correction, by returning non-nil, or reject it, by
returning nil. one use case is that in evil, a vim emulation in emacs, one can
allow esc to accept a speck correction only in insert state.